ajout notes

+++ conteneurisation/src/main/java/project/conteneurisation/repository/NoteRepository.java
@@ -3,4 +3,4 @@ package project.conteneurisation.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import project.conteneurisation.model.Note;
 
-public interface NoteRepository extends JpaRepository<Note, Long> {}
+public interface NoteRepository extends JpaRepository<Note, Integer> {}
